Stop infinite loop when measuring zoom level onresize.
For FF 3.6 and Webkit, I detected zoom by inserting an invisible
element, measuring it, and then removing it. But even though it was
visibility:hidden, it could change the size of the viewport, causing
the "resize" event to be triggered. Since the test-page recalculates
onresize, this caused an infinite loop.

The solution is to put the element to measure inside a new flow (e.g.,
an element with overflow:hidden) which itself has width and height 0.